move: auth components to app dir
