+++ apps/web/app/(auth)/auth/forgot-password/reset/page.tsx
@@ -1,5 +1,5 @@
-import { ResetPasswordForm } from "@/components/auth/ResetPasswordForm";
 import FormWrapper from "@/app/(auth)/auth/FormWrapper";
+import { ResetPasswordForm } from "./ResetPasswordForm";
 
 const ResetPasswordPage: React.FC = () => {
   return (
